finalize assignment 3
