+++ .gitignore
@@ -1,1 +1,3 @@
 *.DS_Store
+LOCAL:*
+__pycache__/
